+++ examples/multisite/example/urls.py
@@ -18,14 +18,8 @@ import os
 from django.conf import settings
 from django.contrib import admin
 from django.urls import include, path
-from wagtail import VERSION as WAGTAIL_VERSION
+from wagtail import urls as wagtail_urls
 from wagtail.admin import urls as wagtailadmin_urls
-
-if WAGTAIL_VERSION >= (3, 0):
-    from wagtail import urls as wagtail_urls
-else:
-    from wagtail.core import urls as wagtail_urls
-
 from wagtail.documents import urls as wagtaildocs_urls
 
 urlpatterns = [