+++ tests/factories/page.py
@@ -1,10 +1,5 @@
 import factory
-from wagtail import VERSION as WAGTAIL_VERSION
-
-if WAGTAIL_VERSION >= (3, 0):
-    from wagtail.models import Page
-else:
-    from wagtail.core.models import Page
+from wagtail.models import Page
 
 
 class PageFactory(factory.DjangoModelFactory):