+++ tests/factories/site.py
@@ -1,10 +1,5 @@
 import factory
-from wagtail import VERSION as WAGTAIL_VERSION
-
-if WAGTAIL_VERSION >= (3, 0):
-    from wagtail.models import Site
-else:
-    from wagtail.core.models import Site
+from wagtail.models import Site
 
 
 class SiteFactory(factory.DjangoModelFactory):