+++ tests/fixtures.py
@@ -1,23 +1,16 @@
 import pytest
-from wagtail import VERSION as WAGTAIL_VERSION
 
 
 @pytest.fixture
 def page():
-    if WAGTAIL_VERSION >= (3, 0):
-        from wagtail.models import Page
-    else:
-        from wagtail.core.models import Page
+    from wagtail.models import Page
 
     return Page.objects.get(slug='home')
 
 
 @pytest.fixture
 def site():
-    if WAGTAIL_VERSION >= (3, 0):
-        from wagtail.models import Site
-    else:
-        from wagtail.core.models import Site
+    from wagtail.models import Site
 
     return Site.objects.get(is_default_site=True)