+++ examples/aws/example/models.py
@@ -1,7 +1,7 @@
-from wagtail.admin.edit_handlers import FieldPanel, StreamFieldPanel
-from wagtail.core import blocks
-from wagtail.core.fields import StreamField
-from wagtail.core.models import Page
+from wagtail import blocks
+from wagtail.admin.panels import FieldPanel
+from wagtail.fields import StreamField
+from wagtail.models import Page
 
 from wagtailbakery.models import AutoPublishingWagtailBakeryModel
 
@@ -13,7 +13,7 @@ class AbstractExamplePage(Page, AutoPublishingWagtailBakeryModel):
 
     content_panels = [
         FieldPanel('title'),
-        StreamFieldPanel('body')
+        FieldPanel('body')
     ]
 
     class Meta: