+++ tests/urls.py
@@ -1,6 +1,6 @@
 from django.urls import include, path
+from wagtail import urls as wagtail_urls
 from wagtail.admin import urls as wagtailadmin_urls
-from wagtail.core import urls as wagtail_urls
 
 urlpatterns = [
     path('admin/', include(wagtailadmin_urls)),