+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/wagtail-nest/wagtail-bakery/compare/0.6.0...HEAD)
+## [Unreleased](https://github.com/wagtail-nest/wagtail-bakery/compare/0.7.0...HEAD)
+
+Nothing yet
+
+## [0.7.0](https://github.com/wagtail-nest/wagtail-bakery/compare/0.6.0...0.7.0)
 
 ### Added